+++ packages/database/src/migrator.ts
@@ -66,6 +66,7 @@ import * as customerRfmProfiles from './migrations/064_customer_rfm_profiles.js'
 import * as membershipRulesAlerts from './migrations/065_membership_rules_alerts.js';
 import * as storeContactQr from './migrations/066_store_contact_qr.js';
 import * as serviceCategoryRank from './migrations/067_service_category_rank.js';
+import * as reviewsReply from './migrations/068_reviews_reply.js';
 import type { Database } from './types.js';
 
 const migrationProvider: MigrationProvider = {
@@ -138,6 +139,7 @@ const migrationProvider: MigrationProvider = {
       '065_membership_rules_alerts': membershipRulesAlerts,
       '066_store_contact_qr': storeContactQr,
       '067_service_category_rank': serviceCategoryRank,
+      '068_reviews_reply': reviewsReply,
     };
   },
 };